legacy: Move wait_random() to common.c and use HMAC DRBG to generate delays.
