+++ legacy/usb_standard.c
@@ -38,6 +38,7 @@ LGPL License Terms @ref lgpl_license
 
 #include <string.h>
 #include <libopencm3/usb/usbd.h>
+#include "common.h"
 #include "usb_private.h"
 #include "util.h"
