+++ legacy/winusb.c
@@ -18,6 +18,7 @@
 
 #include "winusb.h"
 #include <libopencm3/usb/usbd.h>
+#include "common.h"
 #include "util.h"
 
 static int usb_descriptor_type(uint16_t wValue) { return wValue >> 8; }
